use int for max_length and unify name to max_length

+++ fluid/PaddleNLP/neural_machine_translation/transformer/capi/data_reader.cc
@@ -73,7 +73,7 @@ void DataReader::load_dict() {
 bool DataReader::NextBatch(std::vector<std::vector<int64_t>>& inst_data,
                            std::vector<std::vector<int64_t>>& inst_pos,
                            std::vector<std::vector<float>>& slf_attn_bias_data,
-                           size_t& max_length,
+                           int& max_length,
                            int batch_size) {
   // clear vectors
   inst_data.clear();
@@ -100,7 +100,7 @@ bool DataReader::NextBatch(std::vector<std::vector<int64_t>>& inst_data,
     split(line, sentence_sep, &pieces);
     std::vector<int64_t> sentence_indices = convert_to_ind(pieces[0]);
     inst_data[i] = sentence_indices;
-    if (sentence_indices.size() > max_length) {
+    if (sentence_indices.size() > static_cast<size_t>max_length) {
       max_length = sentence_indices.size();
     }
     inst_pos[i].resize(sentence_indices.size());
